blog/requirements.txt
django==1.4.2
django-tastypie